MAIN-T-121 Default search for Documents
Update logic for document retrieval and search pattern

Removed logic that retrieves all documents for a user in `DocumentService` and `DocumentFacade`. Simplified search pattern in `SearchSpec` and `DocumentSearchController`. Adjusted the corresponding API interfaces and modified various configuration files.
