+++ server/.run/server [integrationTest].run.xml
@@ -33,4 +33,38 @@
         <RunAsTest>false</RunAsTest>
         <method v="2"/>
     </configuration>
+    <configuration default="false" name="server [integrationTest]" type="GradleRunConfiguration" factoryName="Gradle">
+        <ExternalSystemSettings>
+            <option name="executionName"/>
+            <option name="externalProjectPath" value="$PROJECT_DIR$/server"/>
+            <option name="externalSystemIdString" value="GRADLE"/>
+            <option name="scriptParameters" value="-PrunIntegrationTests=true"/>
+            <option name="taskDescriptions">
+                <list/>
+            </option>
+            <option name="taskNames">
+                <list>
+                    <option value="integrationTest"/>
+                </list>
+            </option>
+            <option name="vmOptions"/>
+        </ExternalSystemSettings>
+        <ExternalSystemDebugServerProcess>true</ExternalSystemDebugServerProcess>
+        <ExternalSystemReattachDebugProcess>true</ExternalSystemReattachDebugProcess>
+        <EXTENSION ID="com.intellij.execution.ExternalSystemRunConfigurationJavaExtension">
+            <extension name="net.ashald.envfile">
+                <option name="IS_ENABLED" value="false"/>
+                <option name="IS_SUBST" value="false"/>
+                <option name="IS_PATH_MACRO_SUPPORTED" value="false"/>
+                <option name="IS_IGNORE_MISSING_FILES" value="false"/>
+                <option name="IS_ENABLE_EXPERIMENTAL_INTEGRATIONS" value="false"/>
+                <ENTRIES>
+                    <ENTRY IS_ENABLED="true" PARSER="runconfig" IS_EXECUTABLE="false"/>
+                </ENTRIES>
+            </extension>
+        </EXTENSION>
+        <DebugAllEnabled>false</DebugAllEnabled>
+        <RunAsTest>false</RunAsTest>
+        <method v="2"/>
+    </configuration>
 </component>
