+++ server/src/main/kotlin/org/hkurh/doky/documents/db/DocumentEntityRepository.kt
@@ -11,9 +11,6 @@ import java.util.*
  */
 interface DocumentEntityRepository : CrudRepository<DocumentEntity?, Long?>, JpaSpecificationExecutor<DocumentEntity?> {
 
-    @Query("select d from DocumentEntity d where d.creator.id = ?1")
-    fun findByCreatorId(documentId: Long): List<DocumentEntity>
-
     @Query("select d from DocumentEntity d where d.id = ?1 and d.creator.id = ?2")
     fun findByIdAndCreatorId(documentId: Long, creatorId: Long): DocumentEntity?
